requirements.txt
asyncore-wsgi==0.0.11
blinker==1.8.2
bottle==0.13.1
click==8.1.7
colorama==0.4.6
Flask==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
packaging==24.1
python-dotenv==1.0.1
web-pdb==1.6.2
Werkzeug==3.0.4
